LayoutTests/imported/w3c/web-platform-tests/html/dom/elements/global-attributes/cue.html

<!DOCTYPE html>
<html>
    <head>
        <title>Test for cue attribute</title>
        <link rel="help" href="https://github.com/whatwg/html/pull/9771" />
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <span id="element1"></span>
        <span id="element2" cue></span>
        <span id="element3" cue="custom value"></span>
    </body>
    <script>
        const element1 = document.getElementById("element1");
        const element2 = document.getElementById("element2");
        const element3 = document.getElementById("element3");
        test(() => {
            assert_false(element1.cue);
        }, '"cue" attribute in HTML defaults to false');

        test(() => {
            element1.setAttribute("CUE", "");
            assert_true(element1.cue);
            element1.removeAttribute("CUE");
            assert_false(element1.cue);
        }, "Boolean attributes should be case-insensitive");

        test(() => {
            assert_true(element2.cue);
        }, 'Setting "cue" attribute in HTML should set the property to true');

        test(() => {
            assert_true(element3.cue);
            assert_equals(element3.getAttribute("cue"), "custom value");
            element3.setAttribute("cue", "new value")
            assert_true(element3.cue);
            assert_equals(element3.getAttribute("cue"), "new value");
            element3.cue = false;
            assert_equals(element3.getAttribute("cue"), null);
        }, 'Getting a present "cue" attribute should return the attribute value');

        test(() => {
            element1.setAttribute("cue", "");
            assert_true(element1.cue);
        }, 'Setting "cue" attribute using setAttribute should set the property to true');

        test(() => {
            element1.cuebackground = true;
            assert_equals(element1.getAttribute("cue"), "");
        }, "Setting the 'cue' to true should set its value to empty string");

        test(() => {
            element1.cue = true;
            assert_true(element1.cue);
            element1.cue = false;
            assert_false(element1.cue);
            element1.cue = true;
            assert_true(element1.cue);
        }, 'Toggling "cue" attribute should result in the correct value');

        test(() => {
            element1.cue = true;
            const clone = element1.cloneNode(true);
            assert_true(clone.cue);
        }, "Cloning an element should preserve the boolean attribute");

        test(() => {
            const dynamicElement = document.createElement("example-element");
            assert_false(dynamicElement.cue);
            dynamicElement.cue = true;
            assert_true(dynamicElement.cue);
        }, "Boolean attribute should work correctly on dynamically created elements");

        test(() => {
            element1.setAttribute("cue", "");
            assert_true(element1.cue);
            element1.removeAttribute("cue");
            assert_false(element1.cue);
        }, 'Removing "cue" attribute using removeAttribute should set the property to false');

        test(() => {
            element1.cue = true;
            assert_true(element1.cue);
            element1.cue = false;
            assert_false(element1.cue);
        }, 'Setting "cue" property via JavaScript should reflect the correct boolean value');

        test(() => {
            element1.cue = false;
            assert_false(element1.hasAttribute("cue"));
            element1.cue = true;
            assert_true(element1.hasAttribute("cue"));
        }, 'Setting "cue" property via JavaScript should reflect in the content attribute');

        const truthyValues = [true, 1, "true", {}, [], "any-non-empty-string"];
        truthyValues.forEach((value) => {
            test(() => {
                element1.cue = value;
                assert_true(element1.cue);
            }, `Setting "cue" property in JavaScript with truthy value (${typeof value} ${value}) should set the property to true`);
        });

        const falsyValues = [false, 0, "", null, undefined];
        falsyValues.forEach((value, index) => {
            test(() => {
                element1.cue = value;
                assert_false(element1.cue);
            }, `Setting "cue" property in JavaScript with falsy value (${value}) should set the property to false`);
        });
    </script>
</html>
